[#497] Fixed 32 signed integer overflow with RS Bridge functions when too many  external storages(Or with a huge size) or too many disks were used.

+++ src/main/java/de/srendi/advancedperipherals/common/addons/refinedstorage/RefinedStorage.java
@@ -113,8 +113,8 @@ public class RefinedStorage {
         return network.getCraftingManager().getPattern(stack) != null;
     }
 
-    public static int getMaxItemDiskStorage(INetwork network) {
-        int total = 0;
+    public static long getMaxItemDiskStorage(INetwork network) {
+        long total = 0;
         boolean creative = false;
         for (IStorage<ItemStack> store : network.getItemStorageCache().getStorages()) {
             if (store instanceof IStorageDisk<ItemStack> storageDisk) {
@@ -126,8 +126,8 @@ public class RefinedStorage {
         return creative ? -1 : total;
     }
 
-    public static int getMaxFluidDiskStorage(INetwork network) {
-        int total = 0;
+    public static long getMaxFluidDiskStorage(INetwork network) {
+        long total = 0;
         boolean creative = false;
         for (IStorage<FluidStack> store : network.getFluidStorageCache().getStorages()) {
             if (store instanceof IStorageDisk<FluidStack> storageDisk) {
@@ -139,8 +139,8 @@ public class RefinedStorage {
         return creative ? -1 : total;
     }
 
-    public static int getMaxItemExternalStorage(INetwork network) {
-        int total = 0;
+    public static long getMaxItemExternalStorage(INetwork network) {
+        long total = 0;
         for (IStorage<ItemStack> store : network.getItemStorageCache().getStorages()) {
             if (store instanceof IExternalStorage<ItemStack> externalStorage) {
                 total += externalStorage.getCapacity();
@@ -149,8 +149,8 @@ public class RefinedStorage {
         return total;
     }
 
-    public static int getMaxFluidExternalStorage(INetwork network) {
-        int total = 0;
+    public static long getMaxFluidExternalStorage(INetwork network) {
+        long total = 0;
         for (IStorage<FluidStack> store : network.getFluidStorageCache().getStorages()) {
             if (store instanceof IExternalStorage<FluidStack> externalStorage) {
                 total += externalStorage.getCapacity();
